Split Exception and domain
