+++ src/main/java/subway/station/StationService.java
@@ -1,6 +1,5 @@
 package subway.station;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
